bgp: move BGPConfig to pkg/bgp/config
Move BGPConfig out of pkg/bgp/option into a new pkg/bgp/config package,
to decouple BGP CP cell from the common daemon infra.

Rename EnableBGPLegacyOriginAttribute to EnableLegacyOriginAttribute, the
BGP prefix is redundant inside the BGP config struct.

[ mrasek: Rebased on 'main', split out of the original flag move commit ]
